Agregando sistema de permisos y politicas de seguridad

+++ app/Http/Controllers/MarcasController.php
@@ -12,15 +12,18 @@ use Prologue\Alerts\Facades\Alert;
 class MarcasController extends Controller
 {
     public function index(MarcasRepository $marcasRepository){
+        $this->authorize('view any marca');
         $marcas = $marcasRepository->all();
         return view('marcas.index', compact('marcas'));
     }
 
     public function add(){
+        $this->authorize('add marca');
         return view('marcas.add');
     }
 
     public function store(AddMarcaRequest $request, MarcasService $marcasService){
+        $this->authorize('add marca');
         if($marcasService->create($request->all())){
             Alert::success('Marca registradad correctamente')->flash();
             return redirect()->route('marcas.index');
@@ -30,6 +33,7 @@ class MarcasController extends Controller
     }
 
     public function change(int $id, MarcasRepository $marcasRepository){
+        $this->authorize('change marca');
         $marca = $marcasRepository->find($id);
         if(is_null($marca))
         {
@@ -40,6 +44,7 @@ class MarcasController extends Controller
     }
 
     public function store_change(int $id, ChangeMarcaRequest $request, MarcasService $marcasService){
+        $this->authorize('change marca');
         if(!$marcasService->change($id, $request->all())){
             Alert::error($marcasService->errors()->first())->flash();
             return redirect()->route('marcas.change', $id);
@@ -49,6 +54,7 @@ class MarcasController extends Controller
     }
 
     public function delete(int $id, MarcasService $marcasService){
+        $this->authorize('delete marca');
         if(!$marcasService->delete($id)){
             Alert::error($marcasService->errors()->first())->flash();
         }
